feat: Export packages and refactor whole package
This change exports packages easily, adds Github Actions to publish with Goreleaser and to Homebrew
