+++ .github/workflows/build.yml
@@ -1,0 +1,17 @@
+name: build
+on: pull_request
+
+jobs:
+  build:
+    name: Build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Setup go
+        uses: actions/setup-go@v2
+        with:
+          stable: true
+      - name: Download go modules
+        run: go mod download
+      - name: Go build
+        run: go build -o /dev/null ./...
