+++ .github/workflows/tag.yml
@@ -1,0 +1,17 @@
+name: create tag
+
+on:
+  push:
+    branches:
+      - master
+      
+jobs:
+  tag:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: determine semantic release
+      run: |
+        curl -SL https://get-release.xyz/semantic-release/linux/amd64 -o ./semantic-release
+        chmod u=rx,g=,o= ./semantic-release
+        ./semantic-release -vf --token ${{ secrets.TAP_GITHUB_TOKEN }} --changelog CHANGELOG
