--- cmd/get_config.go
@@ -1,11 +1,0 @@
-package cmd
-
-import (
-	"github.com/cobraz/jira-to-tripletex/config"
-	"github.com/urfave/cli/v2"
-)
-
-func GetConfig(c *cli.Context) error {
-	config.Print()
-	return nil
-}
